Feature/api changes

+++ scheduler/src/main/java/com/jagostinho/scheduler_api/service/TaskService.java
@@ -4,15 +4,18 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.server.ResponseStatusException;
 
-import com.jagostinho.scheduler_api.business.Task;
 import com.jagostinho.scheduler_api.dal.TaskRepository;
+import com.jagostinho.scheduler_api.domain.Task;
+
 import javax.persistence.*;
 
 @Service
@@ -20,39 +23,42 @@ public class TaskService {
 
     private final TaskRepository taskRepository;
 
+    // Dependecy Injection
     @Autowired
     public TaskService(TaskRepository taskRepository) {
         this.taskRepository = taskRepository;
     }
 
-    public List<Task> getTasks() {
-        return taskRepository.findAll();
+    public ResponseEntity<List<Task>> getTasks() {
+        return new ResponseEntity<>(taskRepository.findAll(Sort.by(Sort.Direction.DESC, "createdAt")), HttpStatus.OK);
     }
 
-    public Task getTaskById(Long id) {
+    public ResponseEntity<Task> getTaskById(Long id) {
         Optional<Task> optionalTask = taskRepository.findById(id);
         if (optionalTask.isPresent()) {
-            return optionalTask.get();
+            return new ResponseEntity<>(optionalTask.get(), HttpStatus.OK);
         }
-        throw new ResponseStatusException(
-                HttpStatus.NOT_FOUND, String.format("task with the id '%d' not found", id));
+        return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
     }
 
-    public void createTask(Task task) {
-        taskRepository.save(task);
+    public ResponseEntity<Task> createTask(Task task) {
+        return new ResponseEntity<>(taskRepository.save(task), HttpStatus.CREATED);
     }
 
-    public void updateTask(Task task) {
+    public ResponseEntity<Task> updateTask(Task task) {
         Optional<Task> optionalTask = taskRepository.findById(task.getId());
         if (optionalTask.isPresent()) {
-            taskRepository.save(task);
-        } else {
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND, String.format("task with the id '%d' not found", task.getId()));
+            return new ResponseEntity<>(taskRepository.save(task), HttpStatus.OK);
         }
+        return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
     }
 
-    public void deleteTask(Long id) {
-        taskRepository.deleteById(id);
+    public ResponseEntity<Task> deleteTask(Long id) {
+        Optional<Task> optionalTask = taskRepository.findById(id);
+        if (optionalTask.isPresent()) {
+            taskRepository.deleteById(id);
+            return new ResponseEntity<>(null, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
     }
 }
